GIX-1115: Revert changes
