+++ packages/sns/package.json
@@ -38,7 +38,7 @@
   "peerDependencies": {
     "@dfinity/utils": "^0.0.6"
   },
-  "devDependencies": {
-    "jest-environment-jsdom": "^29.3.1"
+  "dependencies": {
+    "js-sha256": "^0.9.0"
   }
 }
